+++ exporters/cert/exporter.go
@@ -13,10 +13,6 @@ import (
 	"github.com/spf13/afero"
 )
 
-const (
-	certType = "file"
-)
-
 type Config struct {
 	Paths []string
 }
@@ -81,7 +77,7 @@ func (e *Exporter) collectPath(ch chan<- prometheus.Metric, path string) error {
 
 			for _, cert := range certs {
 				timestamp := float64(cert.NotAfter.Unix())
-				ch <- prometheus.MustNewConstMetric(e.cert, prometheus.GaugeValue, timestamp, fpath, certType)
+				ch <- prometheus.MustNewConstMetric(e.cert, prometheus.GaugeValue, timestamp, fpath)
 			}
 			e.logger.Log("info", fmt.Sprintf("added %s to the metrics", fpath))
 
@@ -126,7 +122,6 @@ func New(config Config) (*Exporter, error) {
 			"Timestamp after which the cert is invalid.",
 			[]string{
 				"path",
-				"type",
 			},
 			nil,
 		),